requirements.txt
beautifulsoup4
black
tiktoken
openai
unstructured
nltk
fastapi
jinja2
uvicorn
streamlit
streamlit-chat
tqdm
isort
langchainhub
firecrawl-py
langchain-community
langchain-openai
langchain-core
langchain-pinecone
langsmith
pillow